# Unify reader implementations