Use an atomic.Value to manage concurrency without a lock.

Resolves #2921 

### benchstat against `new_sdk/main`

```
name                      old time/op    new time/op    delta
ManualReader/Collect-8      19.3ns ± 1%     7.5ns ± 7%  -61.32%  (p=0.000 n=10+10)
PeriodicReader/Collect-8    6.88ns ± 4%    6.96ns ± 4%     ~     (p=0.315 n=9+10)

name                      old alloc/op   new alloc/op   delta
ManualReader/Collect-8       0.00B          0.00B          ~     (all equal)
PeriodicReader/Collect-8     0.00B          0.00B          ~     (all equal)

name                      old allocs/op  new allocs/op  delta
ManualReader/Collect-8        0.00           0.00          ~     (all equal)
PeriodicReader/Collect-8      0.00           0.00          ~     (all equal)
```